Move TextMagnifierConfiguration to magnifier.dart
